+++ test/PoseidonGoldenTests/GoldenTestsRunCommands.hs
@@ -34,7 +34,8 @@ import           Poseidon.GenotypeData      (GenoDataSource (..),
                                              GenotypeDataSpec (..),
                                              GenotypeFormatSpec (..),
                                              SNPSetSpec (..))
-import           Poseidon.ServerClient      (ArchiveEndpoint (..))
+import           Poseidon.ServerClient      (AddJannoColSpec (..),
+                                             ArchiveEndpoint (..))
 import           Poseidon.Utils             (LogMode (..), TestMode (..),
                                              getChecksum, testLog,
                                              usePoseidonLogger)
@@ -331,7 +332,7 @@ testPipelineList testDir checkFilePath = do
         }
     runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts2) "list" 2
     let listOpts3 = listOpts1 {
-          _listListEntity    = ListIndividuals ["Country", "Nr_SNPs"]
+          _listListEntity    = ListIndividuals (AddJannoColList ["Country", "Nr_SNPs"])
         }
     runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts3) "list" 3
     let listOpts4 = listOpts3 {
@@ -342,6 +343,10 @@ testPipelineList testDir checkFilePath = do
           _listOnlyLatest     = True
         }
     runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts5) "list" 5
+    let listOpts6 = listOpts1 {
+          _listListEntity    = ListIndividuals AddJannoColAll
+        }
+    runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts6) "list" 6
 
 testPipelineSummarise :: FilePath -> FilePath -> IO ()
 testPipelineSummarise testDir checkFilePath = do
@@ -1069,7 +1074,7 @@ testPipelineListRemote testDir checkFilePath = do
             }
         runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts2) "listRemote" 2
         let listOpts3 = listOpts1 {
-              _listListEntity    = ListIndividuals ["Publication"]
+              _listListEntity    = ListIndividuals (AddJannoColList ["Publication"])
             , _listRawOutput     = True
             }
         runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts3) "listRemote" 3
@@ -1082,6 +1087,12 @@ testPipelineListRemote testDir checkFilePath = do
             , _listOnlyLatest   = False
             }
         runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts4) "listRemote" 4
+
+        let listOpts5 = listOpts1 {
+              _listListEntity    = ListIndividuals AddJannoColAll
+            , _listRawOutput     = True
+            }
+        runAndChecksumStdOut checkFilePath testDir (testLog $ runList listOpts5) "listRemote" 5
         ) (
         killThread threadID
         )